+++ src/main/java/ysaak/garde/business/utils/ContractUtils.java
@@ -20,7 +20,7 @@ public final class ContractUtils {
    * @param attendanceWeek Attendance weeks
    * @return Monthly payment
    */
-  public static double calculateMonthlyPayment(double hourPrice, double hoursPerWeek, int attendanceWeek) {
+  public static double calculateMonthlyPayment(double hourPrice, int hoursPerWeek, int attendanceWeek) {
     double payment = hourPrice * hoursPerWeek * attendanceWeek / 12.;
 
     final BigDecimal bd = BigDecimal.valueOf(payment).setScale(2, RoundingMode.HALF_UP);
@@ -33,7 +33,7 @@ public final class ContractUtils {
    * @param attendanceWeek Attendance weeks
    * @return Normal monthly hours
    */
-  public static int calculateNormalMonthlyHours(double hoursPerWeek, int attendanceWeek) {
+  public static int calculateNormalMonthlyHours(int hoursPerWeek, int attendanceWeek) {
     double hours = hoursPerWeek * attendanceWeek / 12.;
 
     final BigDecimal bd = BigDecimal.valueOf(hours).setScale(0, RoundingMode.HALF_UP);
