+++ src/main/java/ysaak/garde/data/contract/ContractDTO.java
@@ -53,7 +53,7 @@ public class ContractDTO {
   /**
    * Hours per week
    */
-  private Double hoursPerWeek;
+  private Integer hoursPerWeek;
 
   /**
    * Base value of an hour
@@ -129,11 +129,11 @@ public class ContractDTO {
     this.attendancePerWeek = attendancePerWeek;
   }
 
-  public Double getHoursPerWeek() {
+  public Integer getHoursPerWeek() {
     return hoursPerWeek;
   }
 
-  public void setHoursPerWeek(Double hoursPerWeek) {
+  public void setHoursPerWeek(Integer hoursPerWeek) {
     this.hoursPerWeek = hoursPerWeek;
   }
